Ahora si, creo que el commit final

+++ src/main/java/src/vehiculos/Camioneta.java
@@ -5,7 +5,7 @@ public class Camioneta extends Vehiculo{
 	private boolean volco;
 	
 	public Camioneta() {cantidadCamioneta++;}
-	public Camioneta(String placa, int puertas, String nombre, float precio, float peso, Fabricante fabricante, boolean puestos) {
+	public Camioneta(String placa, int puertas, String nombre, int precio, int peso, Fabricante fabricante, boolean puestos) {
 		super(placa,puertas,90,nombre,precio,peso,"4X4",fabricante);
 		this.volco=puestos;
 		cantidadCamioneta++;}
